+++ Src/FluentAssertions/CallerIdentification/ParsingState.cs
@@ -1,0 +1,9 @@
+﻿namespace FluentAssertions.CallerIdentification
+{
+    internal enum ParsingState
+    {
+        InProgress,
+        GoToNextSymbol,
+        Done
+    }
+}